code for day 24 part 2
This uses the linear algebra solution described by user DaveBaum at
https://www.reddit.com/r/adventofcode/comments/18pnycy/2023_day_24_solutions

Even with this it was a little tricky to get right, first because his
syntax doesn't differentiate between multiplation and dot
products. Then the numbers are so large that I was experiencing round
off errors in my answer. Switching to float128 seems to have solved
that problem.