Working code to solve day 25
I deleted all the other code I had and just removed the 3 edges I
found using graphviz. I was worried I wouldn't be able to see them,
and that was in fact a problem for the test input, but it was super
obvious for the full and very large graph.

Merry Belated Christmas!